requirements-dev.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.3
coverage==4.5.4
cycler==0.10.0
decorator==4.4.1
docutils==0.16
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
furl==2.1.0
grpcio==1.26.0
grpcio-tools==1.26.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
ipython==7.10.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mock==4.0.2
more-itertools==7.2.0
numpy==1.18.2
orderedmultidict==1.0.1
packaging==19.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.2
pipdeptree==0.13.2
pluggy==0.13.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.3.1
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.15
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
